requirements.txt
Flask_RESTful==0.3.9
Flask==2.0.1
mysql_connector_repackaged==0.3.1
